Refactor AddHackathonFragment to CreateHackthonFragment, also refactor all relevant methods and classes to match name
