+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/FragmentComponent.kt
@@ -2,7 +2,7 @@ package com.lambdaschool.hackathon_portal.di
 
 import androidx.fragment.app.Fragment
 import com.lambdaschool.hackathon_portal.ui.fragments.account.AccountFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.add.AddHackathonFragment
+import com.lambdaschool.hackathon_portal.ui.fragments.create.CreateHackathonFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.dashboard.DashboardFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginModule
@@ -27,5 +27,5 @@ interface FragmentComponent {
     fun injectDashboardFragment(fragment: DashboardFragment)
     fun injectAccountFragment(fragment: AccountFragment)
     fun injectSettingsFragment(fragment: SettingsFragment)
-    fun injectAddHackathonFragment(fragment: AddHackathonFragment)
+    fun injectCreateHackathonFragment(fragment: CreateHackathonFragment)
 }
