+++ app/src/main/java/com/lambdaschool/hackathon_portal/ui/fragments/create/CreateHackathonViewModel.kt
@@ -1,4 +1,4 @@
-package com.lambdaschool.hackathon_portal.ui.fragments.add
+package com.lambdaschool.hackathon_portal.ui.fragments.create
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
@@ -6,7 +6,7 @@ import com.lambdaschool.hackathon_portal.HackathonRepository
 import com.lambdaschool.hackathon_portal.model.Hackathon
 import javax.inject.Inject
 
-class AddHackathonViewModel @Inject constructor(private val repo: HackathonRepository): ViewModel() {
+class CreateHackathonViewModel @Inject constructor(private val repo: HackathonRepository): ViewModel() {
 
     fun postHackathon(hackathon: Hackathon): LiveData<Boolean> {
         return repo.postHackathon(hackathon)
