+++ app/src/main/res/layout/fragment_create_hackathon.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.fragments.add.AddHackathonFragment">
+    tools:context=".ui.fragments.create.CreateHackathonFragment">
 
     <EditText
         android:id="@+id/edit_text_hackathon_name"
