+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -6,6 +6,10 @@
     <group android:checkableBehavior="single">
 
         <item
+            android:id="@+id/nav_drawer_create_hackathon"
+            android:title="@string/create_hackathon"/>
+
+        <item
             android:id="@+id/nav_drawer_account"
             android:title="@string/manage_account" />
 
@@ -16,10 +20,6 @@
         <item
             android:id="@+id/nav_drawer_logout"
             android:title="@string/logout" />
-
-        <item
-            android:id="@+id/nav_drawer_add_hackathon"
-            android:title="Add Hackathon"/>
 
     </group>
